requirements.txt
audioop-lts
discord.py
canvasapi
python-dotenv
